+++ tests/browser/wait_for_request.test.js
@@ -4,10 +4,9 @@ const Wendigo = require('../..');
 const configUrls = require('../config.json').urls;
 const utils = require('../test_utils');
 
-describe.skip("Wait For Request", function() {
+describe("Wait For Request", function() {
     this.timeout(5000);
     let browser;
-
 
     before(async() => {
         browser = await Wendigo.createBrowser();
@@ -25,17 +24,17 @@ describe.skip("Wait For Request", function() {
         setTimeout(() => {
             browser.clickText("click me");
         }, 100);
-        await browser.assert.request.url(/api/).exactly(0);
-        await browser.waitForRequest("http://localhost:3456/api");
-        await browser.assert.request.url(/api/);
+        await browser.assert.requests.url(/api/).exactly(0);
+        await browser.requests.waitForRequest("http://localhost:3456/api");
+        await browser.assert.requests.url(/api/);
     });
 
     it("Wait For Request Already Made", async() => {
         await browser.clickText("click me");
         await browser.wait(10);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.waitForRequest("http://localhost:3456/api");
-        await browser.assert.request.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.requests.waitForRequest("http://localhost:3456/api");
+        await browser.assert.requests.url(/api/).exactly(1);
     });
 
     it("Wait For Request With Mock", async() => {
@@ -43,9 +42,9 @@ describe.skip("Wait For Request", function() {
         setTimeout(() => {
             browser.clickText("click me");
         }, 100);
-        await browser.assert.request.url(/api/).exactly(0);
-        await browser.waitForRequest("http://localhost:3456/api");
-        await browser.assert.request.url(/api/);
+        await browser.assert.requests.url(/api/).exactly(0);
+        await browser.requests.waitForRequest("http://localhost:3456/api");
+        await browser.assert.requests.url(/api/);
     });
 
     it("Wait For Response With Mock", async() => {
@@ -55,10 +54,10 @@ describe.skip("Wait For Request", function() {
         });
         await browser.clickText("click me");
         await browser.wait(10);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(0);
-        await browser.waitForResponse("http://localhost:3456/api", 1000);
-        await browser.assert.request.url(/api/).responseBody("test");
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(0);
+        await browser.requests.waitForResponse("http://localhost:3456/api", 1000);
+        await browser.assert.requests.url(/api/).responseBody("test");
     });
 
     it("Wait For Response Already Made", async() => {
@@ -67,40 +66,39 @@ describe.skip("Wait For Request", function() {
         });
         await browser.clickText("click me");
         await browser.wait(100);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(1);
-        await browser.waitForResponse("http://localhost:3456/api");
-        await browser.assert.request.url(/api/).responseBody("test").exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(1);
+        await browser.requests.waitForResponse("http://localhost:3456/api");
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(1);
     });
 
     it("Wait For Request Timeout", async() => {
         await utils.assertThrowsAsync(async() => {
-            await browser.waitForRequest("http://localhost:3456/api", 10);
+            await browser.requests.waitForRequest("http://localhost:3456/api", 10);
         }, `TimeoutError: [waitForRequest] Waiting for request "http://localhost:3456/api", timeout of 10ms exceeded.`);
     });
 
     it("Wait For Response Timeout", async() => {
         await utils.assertThrowsAsync(async() => {
-            await browser.waitForResponse("http://localhost:3456/api", 10);
+            await browser.requests.waitForResponse("http://localhost:3456/api", 10);
         }, `TimeoutError: [waitForResponse] Waiting for response "http://localhost:3456/api", timeout of 10ms exceeded.`);
     });
-
 
     it("Wait For Next Request", async() => {
         setTimeout(() => {
             browser.clickText("click me");
         }, 100);
-        await browser.assert.request.url(/api/).exactly(0);
-        await browser.waitForNextRequest("http://localhost:3456/api");
-        await browser.assert.request.url(/api/);
+        await browser.assert.requests.url(/api/).exactly(0);
+        await browser.requests.waitForNextRequest("http://localhost:3456/api");
+        await browser.assert.requests.url(/api/);
     });
 
     it("Wait For Next Request Already Made", async() => {
         await browser.clickText("click me");
         await browser.wait(100);
-        await browser.assert.request.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
         await utils.assertThrowsAsync(async() => {
-            await browser.waitForNextRequest("http://localhost:3456/api", 10);
+            await browser.requests.waitForNextRequest("http://localhost:3456/api", 10);
         }, `TimeoutError: [waitForNextRequest] Waiting for request "http://localhost:3456/api", timeout of 10ms exceeded.`);
     });
 
@@ -111,10 +109,10 @@ describe.skip("Wait For Request", function() {
         });
         await browser.clickText("click me");
         await browser.wait(100);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(0);
-        await browser.waitForNextResponse("http://localhost:3456/api", 1000);
-        await browser.assert.request.url(/api/).responseBody("test");
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(0);
+        await browser.requests.waitForNextResponse("http://localhost:3456/api", 1000);
+        await browser.assert.requests.url(/api/).responseBody("test");
     });
 
     it("Wait For Next Response Already Made", async() => {
@@ -123,11 +121,11 @@ describe.skip("Wait For Request", function() {
         });
         await browser.clickText("click me");
         await browser.wait(100);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(1);
         await utils.assertThrowsAsync(async() => {
-            await browser.waitForNextResponse("http://localhost:3456/api", 10);
+            await browser.requests.waitForNextResponse("http://localhost:3456/api", 10);
         }, `TimeoutError: [waitForNextResponse] Waiting for response "http://localhost:3456/api", timeout of 10ms exceeded.`);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(1);
     });
 });